factor out claims module
