+++ app/routes/artefact.$artefactId.quote.$quoteId.tsx
@@ -13,11 +13,11 @@ import {
 	TabsTrigger,
 } from '#app/components/ui/tabs.tsx'
 import { db } from '#app/db.ts'
-import { getArtefact } from '#app/repositories/artefact.ts'
-import { getCandidateClaims } from '#app/repositories/candidate-claim.ts'
+import { getArtefact } from '#app/modules/claims/artefact-repository.ts'
+import { getCandidateClaims } from '#app/modules/claims/candidate-claim-repository.ts'
+import { getQuoteFallacies } from '#app/modules/claims/quote-fallacy-repository.ts'
+import { getQuote } from '#app/modules/claims/quote-repository.ts'
 import { getPollPost } from '#app/repositories/polls.ts'
-import { getQuoteFallacies } from '#app/repositories/quote-fallacy.ts'
-import { getQuote } from '#app/repositories/quote.ts'
 import {
 	type Artefact,
 	PollType,
